+++ app/src/main/java/org/MdmSystemTools/Application/model/repository/MeetingRepository.kt
@@ -6,11 +6,11 @@ import org.MdmSystemTools.Application.model.entity.Meeting
 interface MeetingRepository {
   suspend fun insert(meeting: Meeting): Long
 
-  suspend fun delete(id: Int)
+  suspend fun delete(id: Long)
 
   suspend fun delete(meeting: Meeting)
 
   suspend fun getAll(): Flow<List<Meeting>>
 
-  suspend fun getById(id: Int): Meeting
+  suspend fun getById(id: Long): Meeting
 }
