+++ app/src/main/java/org/MdmSystemTools/Application/model/repository/ProjectRepository.kt
@@ -6,11 +6,11 @@ import org.MdmSystemTools.Application.model.entity.Project
 interface ProjectRepository {
   fun insert(project: Project): Long
 
-  suspend fun delete(id: Int)
+  suspend fun delete(id: Long)
 
   suspend fun delete(project: Project)
 
   fun getAll(): Flow<List<Project>>
 
-  suspend fun getById(id: Int): Project
+  suspend fun getById(id: Long): Project
 }
